--- app/views/authorization_grants/new.html.erb
@@ -1,13 +1,0 @@
-<div class="flex h-screen">
-  <div class="block w-full m-auto">
-    <div class="card">
-      <%= tag.h2 t('.title'), class: 'text-2xl semibold mb-2' %>
-      <%= tag.p t('.lede'), class: 'mb-2' %>
-      <%= tag.p client_name, class: 'text-center text-xl semibold text-blue-800 bg-blue-100 rounded border-1 border-blue-800 p-8' %>
-      <div class="flex justify-between my-4">
-        <%= button_to t('.reject_cta'), authorization_grants_path, params: { state:, approve: false }, class: 'btn-outline' %>
-        <%= button_to t('.approve_cta'), authorization_grants_path, params: { state:, approve: true }, class: 'btn-primary'  %>
-      </div>
-    </div>
-  </div>
-</div>
