+++ app/views/oauth/authorization_grants/new.html.erb
@@ -1,0 +1,13 @@
+<div class="flex h-screen">
+  <div class="block w-full m-auto">
+    <div class="card">
+      <%= tag.h2 t('oauth.authorization_grants.new.title'), class: 'text-2xl semibold mb-2' %>
+      <%= tag.p t('oauth.authorization_grants.new.lede'), class: 'mb-2' %>
+      <%= tag.p client_name, class: 'text-center text-xl semibold text-blue-800 bg-blue-100 rounded border-1 border-blue-800 p-8' %>
+      <div class="flex justify-between my-4">
+        <%= button_to t('oauth.authorization_grants.new.reject_cta'), oauth_authorization_grants_path, params: { state:, approve: false }, class: 'btn-outline' %>
+        <%= button_to t('oauth.authorization_grants.new.approve_cta'), oauth_authorization_grants_path, params: { state:, approve: true }, class: 'btn-primary'  %>
+      </div>
+    </div>
+  </div>
+</div>
